requirements.txt
folium==0.18.0
geopy==2.4.1
haversine==2.9.0
numpy==2.2.2
pandas==2.2.3
Pillow==11.1.0
plotly==5.24.1
plotly_express==0.4.1
streamlit==1.40.2
streamlit_folium==0.23.2
